+++ angular-primeng/src/app/app.ts
@@ -7,4 +7,6 @@ import { RouterOutlet } from '@angular/router';
 	templateUrl: './app.html',
 	changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class AppComponent {}
+export class AppComponent {
+	protected readonly title = 'angular-primeng';
+}